Automata/requirements.txt
paho-mqtt==1.6.1
pip==23.2.1
setuptools==68.0.0
wheel==0.41.1
